pyproject.toml
[project]
name = "oneplk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.5",
]
